api_consultaLLM/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.29
psycopg2-binary==2.9.6
python-dotenv==1.0.0
google-generativeai==0.3.1
langchain==0.0.267
langchain-google-genai==0.0.5
